Fuzzing_engine/requirements.txt
cantools==39.4.11
docx==0.2.4
lauterbach_trace32_rcl==1.1.0
numpy==2.1.3
openpyxl==3.1.5
plotly==5.24.1
python_can==4.4.2
termcolor==2.5.0
xlwings==0.32.2
